requirements.txt
lucide==1.1.3
Django==5.2
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-storages==1.14.6
boto3==1.38.5
django-environ==0.11.2

djangorestframework==3.16.0
# lang learning videos
youtube-transcript-api==1.0.3
openai==1.76.0
pydantic==2.11.3
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0

dotenv==0.9.9

django-guest-user==0.5.5